mychat/requirements.txt
gradio 
transformers 
pandasai 
langchain-groq 
mysql-connector-python 
torch
tabulate